third

+++ Binomial.java
@@ -18,34 +18,40 @@ public class Binomial {
 	}
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
-		if (n < 0 || k < 0) return 0;
-        int[][] memo = new int[n + 1][k + 1];
+	public static long binomial(int n, int k) {
+        if (n < 0 || k < 0) return 0;
+        
+        
+        long[][] memo = new long[n + 1][k + 1];
+        
         for (int i = 0; i <= n; i++) {
             for (int j = 0; j <= k; j++) {
                 memo[i][j] = -1;
             }
         }
         return binomial(n, k, memo);
-	}
+    }
 
-	private static int binomial(int n, int k, int[][] memo) {
-		if (memo[n][k] != -1) {
-			return memo[n][k];
-		}
-		// Base case
-		if ((k > n)) {
-		   	memo[n][k] = 0; 
-		   	return 0;
-		}
-		// Another base case
-		if (n == 0 || k == 0) {
-		   	memo[n][k] = 1; 
-		   	return 1;
-		}
-		memo[n][k] = binomial(n - 1, k, memo) + binomial(n - 1, k - 1, memo);
-		return memo[n][k];
-	}
+    
+    private static long binomial(int n, int k, long[][] memo) {
+        if (memo[n][k] != -1) {
+            return memo[n][k];
+        }
+        // Base case
+        if (k > n) {
+            memo[n][k] = 0; 
+            return 0;
+        }
+        // Another base case
+        if (n == 0 || k == 0) {
+            memo[n][k] = 1; 
+            return 1;
+        }
+        
+        
+        memo[n][k] = binomial(n - 1, k, memo) + binomial(n - 1, k - 1, memo);
+        return memo[n][k];
+    }
 }
